+++ CHANGELOG.md
@@ -5,6 +5,11 @@ See updating [Changelog example here](https://keepachangelog.com/en/1.0.0/)
 
 ## [Unreleased]
 
+### Changed
+- **Breaking**, Managed Database: `ManagedDatabaseUserOpernSearchAccessControl` fields changed to pointers
+- **Breaking**, Managed Database: `ManagedDatabaseUserPGAccessControl` fields changed to pointers
+- **Breaking**, Managed Database: `ManagedDatabaseUserRedisAccessControl` fields changed to pointers
+
 ### Removed
 - **Breaking**, Managed Database: connection related methods in favor of session